requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cssselect==1.2.0
exceptiongroup==1.0.4
h11==0.14.0
html-similarity==0.3.3
idna==3.4
lxml==4.9.1
outcome==1.2.0
parsel==1.6.0
PySocks==1.7.1
selenium==4.7.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
w3lib==2.1.0
wsproto==1.2.0
xmltodict==0.13.0

Pillow~=9.3.0
ImageHash~=4.3.1